requirements.txt
asttokens==2.4.1
attrs==23.2.0
<<<<<<< HEAD
beautifulsoup4==4.12.3
bleach==6.1.0
=======
>>>>>>> image_processing
certifi==2024.2.2
charset-normalizer==3.3.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
<<<<<<< HEAD
defusedxml==0.7.1
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
=======
exceptiongroup==1.2.1
executing==2.0.1
>>>>>>> image_processing
flatbuffers==24.3.25
fonttools==4.52.1
humanfriendly==10.0
idna==3.7
imageio==2.34.1
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
<<<<<<< HEAD
Jinja2==3.1.4
=======
>>>>>>> image_processing
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
<<<<<<< HEAD
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistune==3.0.2
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
=======
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
matplotlib==3.9.0
matplotlib-inline==0.1.7
mpmath==1.3.0
>>>>>>> image_processing
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
onnxruntime==1.18.0
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
<<<<<<< HEAD
pandocfilters==1.5.1
=======
>>>>>>> image_processing
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
pooch==1.8.1
prompt-toolkit==3.0.43
protobuf==5.27.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==26.0.3
referencing==0.35.1
rembg==2.0.57
requests==2.32.2
rpds-py==0.18.1
scikit-image==0.23.2
scipy==1.13.1
six==1.16.0
<<<<<<< HEAD
soupsieve==2.5
stack-data==0.6.3
sympy==1.12
tifffile==2024.5.22
tinycss2==1.3.0
=======
stack-data==0.6.3
sympy==1.12
tifffile==2024.5.22
>>>>>>> image_processing
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
<<<<<<< HEAD
webencodings==0.5.1
=======
>>>>>>> image_processing
